pyproject.toml
[tool.poetry]
name = "ai_assistant"
version = "0.0.2"
authors = [
    "Anand <anand1911991@outlook.com>",
]
description = "Template repo for building GenAI applications"
readme = "README.md"
repository = "https://github.com/Anand191/terraform_tutorial.git"
packages = [
    {include = "ai_assistant"}
]
exclude = ["tests*"]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
google-cloud-discoveryengine = "^0.11.3"
googleapis-common-protos = "^1.60.0"
google-auth = "^2.23.4"
vertexai = "^1.46.0"
pandas = "^2.1.3"
pandera = "^0.18.3"
numpy = "^1.26.2"
loguru = "^0.7.0"
pydantic = "^2.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.0"
functions-framework = "3.3.0"
chromadb = "^0.4.24"
sentence-transformers = "^2.4.0"
transformers = "^4.40.1"
datasets = "^2.14.7"
langchain = "^0.1.16"
langchain-core = "^0.1.46"
langchain-google-vertexai = "^1.0.2"
langchain-community = "^0.0.34"
dspy-ai = "^2.4.5"
geopy = "^2.4.1"
ragas = "^0.1.7"
openpyxl = "^3.1.2" # used by the logging library
typer = "<0.10.0"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
redis = "^5.0.4"
tqdm = "^4.66.2"
google-cloud-documentai = "^2.27.0"
pypdf2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-watch = "^4.2.0"
pylint = "^2.8.3"
pytest-cov = "^2.11.1"
pytest-xdist = "^2.1.0"
joblib = "^1.3.2"
httpx = "^0.24.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
deptry = "^0.16.1"
ruff = "^0.3.5"
pytest-mock = "^3.14.0"

[tool.poetry.group.experimentation.dependencies]
jupyterlab = "^4.0.8"
tabulate = "^0.9.0"
ipywidgets = "^8.1.1"
google-cloud-aiplatform = "^1.29.0"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
spacy = "^3.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers -m 'not integration and not redis'"
markers = [
    "integration: mark test as integration test" ,
    "redis: mark test as redis based"
]

[tool.deptry]
ignore = ["DEP004"]
ignore_notebooks = true
extend_exclude = ["src"]

[tool.deptry.per_rule_ignores]
DEP002 = ["openpyxl"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "models",
    "data",
    "notebooks",
    "logs",
    "assets",
    # this is only temporary
    "src"
]

line-length = 110
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = ["E731"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
